import

+++ Java_Programming/src/com/ritubrata/various/ByteComparison.java
@@ -2,8 +2,8 @@ package com.ritubrata.various;
 
 public class ByteComparison {
 	public static void main(final String[] args) {
-		final Byte b1 = new Byte(10);
-		final Byte b2 = new Byte(10);
+		final Byte b1 = new Byte((byte) 10);
+		final Byte b2 = new Byte((byte) 10);
 		System.out.println(b1==b2);
 	}
